Create ddos
Attention! DDoS attacks are illegal!!!!

This Python script is written for educational and scientific purposes. 
If you will use it for other purposes, all consequences, such as information system failures that will arise from the operation of the program, may cause you to be punished by both national and international jurisdictions.

Regardless; including whether you are going to use this program for educational or scientific purposes, you must use a virtual machine and obtain written consent from the authority of the target system.